Moved `components/Board.tsx` to `util/weightedDistribution.ts` since it only contained `WeightedDistribution`.
