+++ src/util/weightedDistribution.ts
